Proposed Wireless Sensor Network Design for Monitoring Water Quality in Irish Reservoirs
With the introduction of water charges citizens are questioning the
quality of Irish drinking water. Irish Water wants to reassure the
public that Irish drinking water reservoirs are safe. The company is
looking to implement a service where the reservoirs can be monitored by
the company and also by the public through a website. The following
proposal will be an outlay of the best implementation of a wireless
sensor network for monitoring water quality. The sensors must be able
to measure pH, light, oxygen levels and have the capability to float.
Irish Water want to have the most up to date information on any drop in
water quality and as per specification need to monitor water quality in
fifteen minute intervals.